chore(Console): Refractored printing to console to be part of report module.
