+++ src/krisi/report/console.py
@@ -1,0 +1,69 @@
+from typing import TYPE_CHECKING, List, Union
+
+from rich import box
+from rich.console import Group
+from rich.layout import Layout
+from rich.panel import Panel
+
+from krisi.utils.iterable_helpers import group_by_categories
+from krisi.utils.printing import (
+    create_metric_table,
+    create_y_pred_table,
+    metrics_empty_in_category,
+)
+
+if TYPE_CHECKING:
+    from krisi.evaluate.metric import Metric
+    from krisi.evaluate.scorecard import ScoreCard
+
+from krisi.utils.printing import bold, get_term_size
+
+
+def print_metric(obj: "Metric", repr: bool = False) -> str:
+    hyperparams = ""
+    if obj.parameters is not None:
+        hyperparams += "".join(
+            [f"{key} - {value}" for key, value in obj.parameters.items()]
+        )
+    if obj.result is None and obj.result_rolling is not None:
+        result_ = (
+            "[" + ", ".join([f"{result:<0.5}" for result in obj.result_rolling]) + "]"
+        )
+    else:
+        result_ = f"{obj.result:<15.5}"
+
+    return f"{obj.name:>40s} ({obj.key}): {result_}{hyperparams:>15s}"
+
+
+def get_minimal_summary(obj: "ScoreCard") -> str:
+    return f"\n".join(
+        [
+            f"{metric.name:>40s} - {metric.result:<15.5}"
+            for metric in obj.get_all_metrics()
+            if isinstance(metric.result, (float, int))
+        ]
+    )
+
+
+def get_summary(
+    obj: "ScoreCard", categories: List[str], repr: bool = True, with_info: bool = False
+) -> Union[Panel, Layout]:
+
+    category_groups = group_by_categories(list(vars(obj).values()), categories)
+
+    metric_tables = Group(
+        *[create_y_pred_table(obj.classification, obj.y, obj.predictions)]
+        + [
+            create_metric_table(
+                f"{category if category is not None else 'Unknown Category':>15s}",
+                metrics,
+                with_info,
+                show_header=True if index == 0 else False,
+            )
+            for index, (category, metrics) in enumerate(category_groups.items())
+            if not metrics_empty_in_category(metrics)
+        ],
+    )
+
+    title = f"Result of {obj.model_name if repr else bold(obj.model_name)} on {obj.dataset_name if repr else bold(obj.dataset_name)} tested on {obj.sample_type.value if repr else bold(obj.sample_type.value)}"
+    return Panel(metric_tables, title=title, padding=1, box=box.HEAVY_EDGE, expand=True)
