+++ src/krisi/utils/printing.py
@@ -15,8 +15,13 @@ from rich.pretty import Pretty
 from rich.table import Table
 
 from krisi.evaluate.type import MetricCategories, Predictions, SaveModes, Targets
+from krisi.report.library.console.diagrams import (
+    distribution_plot,
+    histogram_plot,
+    line_plot,
+)
 from krisi.utils.console_plot import plotextMixin
-from krisi.utils.iterable_helpers import group_by_categories, isiterable
+from krisi.utils.iterable_helpers import calculate_nans, group_by_categories, isiterable
 
 if TYPE_CHECKING:
     from krisi.evaluate.metric import Metric
@@ -30,51 +35,6 @@ def bold(text: str, rich: bool = True) -> str:
 def get_term_size() -> int:
     term_size = os.get_terminal_size()
     return term_size.columns
-
-
-def iterative_length(obj: Iterable) -> List[int]:
-    object_shape = []
-    num_obj = 0
-    for el in obj:
-        if isinstance(el, Iterable) and not isinstance(el, str):
-            object_shape.append(iterative_length(el))
-        else:
-            num_obj += 1
-    if num_obj > 0:
-        object_shape.append(num_obj)
-    return object_shape
-
-
-def histogram_plot(data, width, height, title):
-    plx.clf()
-    plx.hist(data, 40)
-    plx.plotsize(width, 10)
-    plx.theme("dark")
-
-    return plx.build()
-
-
-def line_plot(data, width, height, title):
-    plx.clf()
-    plx.plot(data, marker="hd")
-    plx.plotsize(width, 10)
-    plx.theme("dark")
-
-    return plx.build()
-
-
-def distribution_plot(data: pd.Series, width: float, height, title: str):
-    plx.clf()
-    data = data.to_frame("values")
-
-    groups = data.groupby("values").size()
-    plx.simple_bar(
-        groups.index,
-        groups.to_list(),
-        width=width,
-    )
-
-    return plx.build()
 
 
 def __display_result(metric: "Metric") -> Union[Pretty, plotextMixin]:
@@ -112,7 +72,7 @@ def __create_metric(metric: "Metric", with_info: bool) -> List[str]:
     return metric_summarized
 
 
-def __create_metric_table(
+def create_metric_table(
     title: str, metrics: List["Metric"], with_info: bool, show_header: bool = False
 ) -> Table:
     table = Table(
@@ -142,7 +102,7 @@ def __create_metric_table(
     return table
 
 
-def __metrics_empty_in_category(metrics: List["Metric"]) -> bool:
+def metrics_empty_in_category(metrics: List["Metric"]) -> bool:
     return (
         metrics is None
         or len(metrics) < 1
@@ -153,13 +113,7 @@ def __metrics_empty_in_category(metrics: List["Metric"]) -> bool:
     )
 
 
-def calculate_nans(ds: Union[Targets, Predictions]) -> int:
-    return ds.isna().sum() if isinstance(ds, pd.Series) else sum(np.isnan(ds))
-
-
-def __create_y_pred_table(
-    classification: bool, y: Targets, preds: Predictions
-) -> Table:
+def create_y_pred_table(classification: bool, y: Targets, preds: Predictions) -> Table:
     if isinstance(y, np.ndarray):
         y = pd.Series(y)
     if isinstance(preds, np.ndarray):
@@ -211,93 +165,3 @@ def __create_y_pred_table(
         f"Start: {str(preds.index[0])}\nEnd: {str(preds.index[-1])}",
     )
     return table
-
-
-def get_summary(
-    obj: "ScoreCard", categories: List[str], repr: bool = True, with_info: bool = False
-) -> Union[Panel, Layout]:
-
-    category_groups = group_by_categories(list(vars(obj).values()), categories)
-
-    metric_tables = Group(
-        *[__create_y_pred_table(obj.classification, obj.y, obj.predictions)]
-        + [
-            __create_metric_table(
-                f"{category if category is not None else 'Unknown Category':>15s}",
-                metrics,
-                with_info,
-                show_header=True if index == 0 else False,
-            )
-            for index, (category, metrics) in enumerate(category_groups.items())
-            if not __metrics_empty_in_category(metrics)
-        ],
-    )
-
-    title = f"Result of {obj.model_name if repr else bold(obj.model_name)} on {obj.dataset_name if repr else bold(obj.dataset_name)} tested on {obj.sample_type.value if repr else bold(obj.sample_type.value)}"
-    return Panel(metric_tables, title=title, padding=1, box=box.HEAVY_EDGE, expand=True)
-
-
-def get_minimal_summary(obj: "ScoreCard") -> str:
-    return f"\n".join(
-        [
-            f"{metric.name:>40s} - {metric.result:<15.5}"
-            for metric in obj.get_all_metrics()
-            if isinstance(metric.result, (float, int))
-        ]
-    )
-
-
-def print_metric(obj: "Metric", repr: bool = False) -> str:
-    hyperparams = ""
-    if obj.parameters is not None:
-        hyperparams += "".join(
-            [f"{key} - {value}" for key, value in obj.parameters.items()]
-        )
-    if obj.result is None and obj.result_rolling is not None:
-        result_ = (
-            "[" + ", ".join([f"{result:<0.5}" for result in obj.result_rolling]) + "]"
-        )
-    else:
-        result_ = f"{obj.result:<15.5}"
-
-    return f"{obj.name:>40s} ({obj.key}): {result_}{hyperparams:>15s}"
-
-
-def save_object(obj: "ScoreCard", path: str) -> None:
-    import dill
-
-    with open(f"{path}/scorecard.pickle", "wb") as file:
-        dill.dump(obj, file)
-
-
-def save_console(
-    obj: "ScoreCard",
-    path: str,
-    with_info: bool,
-    save_modes: List[Union[SaveModes, str]],
-) -> None:
-
-    summary = get_summary(
-        obj,
-        repr=True,
-        categories=[el.value for el in MetricCategories],
-        with_info=with_info,
-    )
-
-    console = Console(record=True, width=120)
-    with console.capture() as capture:
-        console.print(summary)
-
-    if SaveModes.text in save_modes or SaveModes.text.value in save_modes:
-        console.save_text(f"{path}/console.txt", clear=False)
-    if SaveModes.html in save_modes or SaveModes.html.value in save_modes:
-        console.save_html(f"{path}/console.html", clear=False)
-    if SaveModes.svg in save_modes or SaveModes.svg.value in save_modes:
-        console.save_svg(f"{path}/console.svg", title="save_table_svg.py", clear=False)
-
-
-def save_minimal_summary(obj: "ScoreCard", path: str) -> None:
-    text_summary = get_minimal_summary(obj)
-
-    with open(f"{path}/minimal.txt", "w") as f:
-        f.write(text_summary)
